dev-requirements.txt
# 开发依赖 - 仅开发时需要的工具
-r requirements.txt

# 测试工具
pytest>=8.0.0
pytest-cov>=4.1.0

# 代码质量工具
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
ruff>=0.1.9
mypy>=1.3.0

# 安全检查
bandit>=1.7.5

# 文档工具
mkdocs>=1.5.0
mkdocs-material>=9.4.0
codespell>=2.2.6

# Git工具
pre-commit>=3.3.1
conventional-pre-commit>=3.0.0

# 类型检查支持
types-requests>=2.31.0
pandas-stubs>=2.0.0 